% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{build_network}
\alias{build_network}
\title{Build a road network graph from sf LINESTRING data}
\usage{
build_network(roads_sf, crs_out = 3857)
}
\arguments{
\item{roads_sf}{An sf object with LINESTRING geometry}

\item{crs_out}{Integer EPSG code for projected CRS}
}
\value{
A list with components:
\itemize{
  \item roads: cleaned sf object
  \item nodes: sf POINT object with node_id
  \item edges: sf LINESTRING object with from, to, length
  \item graph: igraph object
  \item A: sparse adjacency matrix
}
}
\description{
Build a road network graph from sf LINESTRING data
}
