% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{degree_matrix}
\alias{degree_matrix}
\title{Degree matrix from adjacency}
\usage{
degree_matrix(A)
}
\arguments{
\item{A}{Sparse symmetric adjacency matrix.}
}
\value{
Diagonal sparse matrix of row sums.
}
\description{
Degree matrix from adjacency
}
\keyword{internal}
