% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{.extract_gaussian_draws}
\alias{.extract_gaussian_draws}
\title{Extract draws from a base fit object (adapter)
EDIT FIELDS LATER}
\usage{
.extract_gaussian_draws(base_fit)
}
\description{
Expected:
- x_draws: matrix S x n
- beta_draws: matrix S x p (optional; if absent we treat X beta = 0)
}
\keyword{internal}
