% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_precision.R
\name{icar_sum_to_zero}
\alias{icar_sum_to_zero}
\title{Apply sum-to-zero constraint to ICAR precision}
\usage{
icar_sum_to_zero(Q)
}
\arguments{
\item{Q}{ICAR precision matrix}
}
\value{
Constrained precision matrix
}
\description{
Projects onto the subspace orthogonal to the constant vector.
}
