% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_roads.R
\name{load_roads}
\alias{load_roads}
\title{Load road geometries from sf object or file}
\usage{
load_roads(x)
}
\arguments{
\item{x}{sf object or file path}
}
\description{
Supports sf objects, .rda files containing a single sf object,
or GeoJSON files.
}
\keyword{internal}
