% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-interactive.R
\name{map_roads_interactive_layers}
\alias{map_roads_interactive_layers}
\title{Interactive map with multiple road layers}
\usage{
map_roads_interactive_layers(sf_aug, layers)
}
\arguments{
\item{sf_aug}{sf object with road geometries}

\item{layers}{named character vector:
names = layer labels, values = column names}
}
\value{
leaflet widget
}
\description{
Interactive map with multiple road layers
}
