% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_moran.R
\name{moran_residuals}
\alias{moran_residuals}
\title{Moran's I for trafficCAR residuals}
\usage{
moran_residuals(
  fit,
  type = c("raw", "structured", "unstructured"),
  nsim = 199,
  method = c("analytic", "permutation")
)
}
\arguments{
\item{fit}{A `traffic_fit` object.}

\item{type}{Residual type: "raw" or "unstructured".}

\item{nsim}{Number of permutations for permutation test.}

\item{method}{"analytic" or "permutation".}
}
\value{
An object of class `traffic_moran`.
}
\description{
Computes Moran's I statistic for model residuals using the model adjacency.
}
