% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_observed_fitted}
\alias{plot_observed_fitted}
\title{Plot observed vs predicted traffic values}
\usage{
plot_observed_fitted(fit, data)
}
\arguments{
\item{fit}{traffic_fit}

\item{data}{data.frame}
}
\value{
ggplot
}
\description{
Plot observed vs predicted traffic values
}
