% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spatial.R
\name{plot_predicted}
\alias{plot_predicted}
\title{Plot predicted traffic outcome on road network}
\usage{
plot_predicted(fit, roads)
}
\arguments{
\item{fit}{traffic_fit}

\item{roads}{sf with segment_id}
}
\value{
ggplot
}
\description{
Plot predicted traffic outcome on road network
}
