% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spatial.R
\name{plot_relative_congestion}
\alias{plot_relative_congestion}
\title{Plot relative congestion on road network}
\usage{
plot_relative_congestion(fit, roads)
}
\arguments{
\item{fit}{traffic_fit}

\item{roads}{sf}
}
\value{
ggplot
}
\description{
Shows systematic deviations after accounting for covariates.
}
