% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_roads_static}
\alias{plot_roads_static}
\title{Static map of road-segment values}
\usage{
plot_roads_static(sf_aug, value_col)
}
\arguments{
\item{sf_aug}{sf object with road geometries}

\item{value_col}{numeric column to plot}
}
\value{
ggplot object
}
\description{
Static map of road-segment values
}
