% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rmvnorm_prec}
\alias{rmvnorm_prec}
\title{Draw from a multivariate normal with sparse precision}
\usage{
rmvnorm_prec(Q, b = NULL)
}
\arguments{
\item{Q}{Sparse symmetric positive definite precision matrix (dgCMatrix).}

\item{b}{Optional numeric vector. If NULL, mean is 0.}
}
\value{
Numeric vector x.
}
\description{
Sample \deqn{x \sim N(Q^{-1} b, Q^{-1})} where \eqn{Q} is symmetric positive definite.
}
\keyword{internal}
