% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-roads.R
\docType{data}
\name{roads_datasets}
\alias{roads_datasets}
\alias{roads_small}
\alias{roads_midsize}
\alias{roads_cstat}
\title{Example road network datasets}
\format{
An object of class \code{sf}.
}
\description{
A collection of example road network datasets provided as `sf` objects.
These datasets are intended for demonstration, testing, and benchmarking.
}
\details{
Each dataset contains LINESTRING road geometries suitable for use with
\code{roads_to_segments()} and related functions.


Included datasets:
\describe{
  \item{roads_small}{Small example road network.}
  \item{roads_midsize}{Medium-sized example road network.}
  \item{roads_cstat}{Road network for College Station, TX.}
}
}
\keyword{datasets}
