% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{row_standardize_weights}
\alias{row_standardize_weights}
\title{Row-standardize adjacency matrix}
\usage{
row_standardize_weights(A, zero_policy = c("keep", "error"))
}
\arguments{
\item{A}{Sparse adjacency matrix.}

\item{zero_policy}{What to do with zero-degree nodes.}
}
\value{
Sparse row-standardized weight matrix.
}
\description{
Row-standardize adjacency matrix
}
\keyword{internal}
