% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights_from_adjacency}
\alias{weights_from_adjacency}
\title{Construct spatial weights matrix}
\usage{
weights_from_adjacency(
  A,
  style = c("binary", "row-standardized"),
  symmetrize = FALSE
)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{style}{One of "binary" or "row-standardized".}

\item{symmetrize}{Passed to adjacency coercion.}
}
\value{
Sparse weight matrix.
}
\description{
Construct spatial weights matrix
}
