% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-interactive.R
\name{map_roads_interactive}
\alias{map_roads_interactive}
\title{Interactive map of road-segment traffic measures}
\usage{
map_roads_interactive(
  sf_aug,
  value = c("predicted_speed", "predicted_volume", "relative_congestion"),
  engine = "leaflet"
)
}
\arguments{
\item{sf_aug}{An `sf` object returned by `augment_roads()`.}

\item{value}{Character scalar. One of:
`"predicted_speed"`, `"predicted_volume"`, `"relative_congestion"`.}

\item{engine}{Currently only `"leaflet"` is supported.}
}
\value{
A leaflet widget.
}
\description{
Displays standard traffic quantities such as predicted speed,
predicted volume, or relative congestion on an interactive map.
}
