% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-interactive.R
\name{map_roads_interactive_layers}
\alias{map_roads_interactive_layers}
\title{Interactive map with multiple standard traffic layers}
\usage{
map_roads_interactive_layers(
  sf_aug,
  values = c("predicted_speed", "relative_congestion")
)
}
\arguments{
\item{sf_aug}{sf object with road geometries}

\item{values}{Character vector of traffic measures to include.}
}
\value{
leaflet widget
}
\description{
Interactive map with multiple standard traffic layers
}
