% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_static.R
\name{plot_roads_static}
\alias{plot_roads_static}
\title{Static map of road-segment traffic measures}
\usage{
plot_roads_static(
  sf_aug,
  value = c("predicted_speed", "predicted_volume", "relative_congestion")
)
}
\arguments{
\item{sf_aug}{sf object with road geometries}

\item{value}{One of "predicted_speed", "predicted_volume",
or "relative_congestion"}
}
\value{
ggplot object
}
\description{
Static map of road-segment traffic measures
}
