% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{plot_traffic_map}
\alias{plot_traffic_map}
\title{Quick map helper for augmented roads}
\usage{
plot_traffic_map(roads_aug, fill = c("predicted_mean", "relative_congestion"))
}
\arguments{
\item{roads_aug}{sf object returned by augment_roads()}

\item{fill}{one of "predicted_mean" or "relative_congestion"}
}
\description{
Quick map helper for augmented roads
}
