% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-roads.R
\docType{data}
\name{roads_datasets}
\alias{roads_datasets}
\title{Example road network datasets}
\format{
An object of class `sf` with LINESTRING geometries.
}
\description{
A collection of example road network datasets provided as `sf` objects.
These datasets are intended for demonstration, testing, and benchmarking.
}
\details{
Each object contains LINESTRING road geometries suitable for use with
`roads_to_segments()` and related functions.


Included datasets:
\describe{
  \item{roads_small}{Small road network example.}
  \item{roads_midsize}{Medium-sized road network example.}
  \item{roads_madison}{Road network for Madison, WI.}
}
}
\keyword{datasets}
