% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficCAR-package.R
\docType{package}
\name{trafficCAR-package}
\alias{trafficCAR}
\alias{trafficCAR-package}
\title{trafficCAR: Conditional Autoregressive (CAR) Models for Traffic Flow on Road Networks}
\description{
Tools for simulating and modeling traffic flow on road networks using spatial conditional autoregressive (CAR) models. The package represents road systems as graphs derived from OpenStreetMap data and supports network-based spatial dependence, basic preprocessing, and visualization for spatio-temporal traffic analysis.
}
\author{
\strong{Maintainer}: Madison Ell \email{mell@stat.tamu.edu}

}
\keyword{internal}
